requirements.txt
feedparser==6.0.11
openai
pydub==0.25.1
Requests==2.32.3
SpeechRecognition==3.10.4
streamlit==1.36.0
python-dotenv
pocketsphinx
youtube_dl